vulkan: replace DescriptorBufferCache by generic cache
